news/backend/requirements.txt
requests
beautifulsoup4
playwright
flask
flask-cors
python-dotenv
openai
